requirements.txt
fastapi
uvicorn
numpy
pillow
pydantic
python-multipart